Refactor in factory pattern
