+++ wcps_auth/handlers/__init__.py
@@ -1,0 +1,2 @@
+from .handler_factory import get_handler_for_packet
+
